Added os-release parser and tests
The fields are defaulted according to the man page.
I don't differentiate "Not implemented" and "optional", so it is easier to
declare schema for it on client side.

The difference would be that "not implemented" fields are not present, but
optional fields would be present with None (null or nil).